requirements.txt
doctr==1.9.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
python-docx==0.8.11
openpyxl==3.1.2
Pillow==10.0.0
datetime
reportlab==4.0.4
Flask
Werkzeug
gunicorn 
PyJWT==2.8.0
flask-cors==4.0.0